+++ starkbank/boletopayment/__boletopayment.py
@@ -49,7 +49,7 @@ def create(payments, user=None):
     ## Parameters (required):
     - payments [list of BoletoPayment objects]: list of BoletoPayment objects to be created in the API
     ## Parameters (optional):
-    - user [Project object]: Project object. Not necessary if starkbank.user was set before function call
+    - user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
     ## Return:
     - list of BoletoPayment objects with updated attributes
     """
@@ -62,7 +62,7 @@ def get(id, user=None):
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):
-    - user [Project object]: Project object. Not necessary if starkbank.user was set before function call
+    - user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
     ## Return:
     - BoletoPayment object with updated attributes
     """
@@ -76,7 +76,7 @@ def pdf(id, user=None):
     ## Parameters (required):
     - id [string]: object unique id. ex: "5656565656565656"
     ## Parameters (optional):
-    - user [Project object]: Project object. Not necessary if starkbank.user was set before function call
+    - user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
     ## Return:
     - BoletoPayment pdf file
     """
@@ -115,7 +115,7 @@ def delete(id, user=None):
     ## Parameters (required):
     - id [string]: BoletoPayment unique id. ex: "5656565656565656"
     ## Parameters (optional):
-    - user [Project object]: Project object. Not necessary if starkbank.user was set before function call
+    - user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
     ## Return:
     - deleted BoletoPayment object
     """