+++ tests/utils/webhook.py
@@ -4,6 +4,6 @@ from uuid import uuid4
 
 def generateExampleWebhook():
     return starkbank.Webhook(
-        url=r"https://webhook.site/{uuid}".format(uuid=str(uuid4())),
-        subscriptions=["transfer", "boleto", "boleto-payment", "boleto-holmes", "brcode-payment", "utility-payment", "deposit", "invoice"]
+        url="https://webhook.site/{uuid}".format(uuid=str(uuid4())),
+        subscriptions=["transfer", "boleto", "boleto-payment", "boleto-holmes", "brcode-payment", "utility-payment", "deposit", "invoice"],
     )