+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Organization user
+- Workspace resource
 
 ## [2.2.2] - 2020-11-24
 ### Added